break down sorting into independent function
